# docs/swagger.yaml
basePath: /v1
definitions:
  Exchange_internal_model.ErrorResponseStruct:
    properties:
      code:
        type: integer
      error:
        description: پیام فنی (developer-friendly)
        type: string
      error_code:
        description: کد خطای اختصاصی (مثلاً validation_error)
        type: string
      message:
        description: پیام کاربر (user-friendly)
        type: string
      success:
        type: boolean
    type: object
  Exchange_internal_model.LoginRequest:
    properties:
      captcha_ans:
        example: aBc123
        type: string
      captcha_id:
        example: 9f1b77d3-cb13-4a72-8a64-28de5f82a5c2
        type: string
      email:
        example: user@example.com
        type: string
      password:
        example: P@ssw0rd123
        type: string
    required:
    - captcha_ans
    - captcha_id
    - email
    - password
    type: object
  Exchange_internal_model.RefreshTokenRequest:
    properties:
      refresh_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    required:
    - refresh_token
    type: object
  Exchange_internal_model.RegisterRequest:
    properties:
      captcha_ans:
        example: aBc123
        type: string
      captcha_id:
        example: 9f1b77d3-cb13-4a72-8a64-28de5f82a5c2
        type: string
      email:
        example: user@example.com
        type: string
      full_name:
        example: Ali Rezaei
        type: string
      password:
        example: P@ssw0rd123
        minLength: 8
        type: string
    required:
    - captcha_ans
    - captcha_id
    - email
    - full_name
    - password
    type: object
  Exchange_internal_model.Response:
    properties:
      code:
        description: HTTP status code
        type: integer
      data:
        description: داده (در صورت موفقیت)
      error_code:
        description: کد یکتا برای خطا (اختیاری)
        type: string
      message:
        description: پیام (موفق یا خطا)
        type: string
      success:
        description: وضعیت موفقیت
        type: boolean
    type: object
...
host: localhost:7000
info:
  contact: {}
  description: Exchange cryptocurrency trading platform
  title: Exchange API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: احراز هویت کاربر با ایمیل، رمز عبور و کپچا. در صورت فعال بودن 2FA،
        ارسال کد تایید دو عاملی
      parameters:
      - description: اطلاعات ورود همراه کپچا
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/Exchange_internal_model.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: ورود موفق بدون 2FA
          schema:
            $ref: '#/definitions/Exchange_internal_model.Response'
        "202":
          description: کد 2FA ارسال شد، منتظر تایید باشید
          schema:
            $ref: '#/definitions/Exchange_internal_model.Response'
        "400":
          description: بدنه درخواست یا کپچا نامعتبر
          schema:
            $ref: '#/definitions/Exchange_internal_model.ErrorResponseStruct'
        "401":
          description: اطلاعات ورود یا کپچا اشتباه
          schema:
            $ref: '#/definitions/Exchange_internal_model.ErrorResponseStruct'
        "500":
          description: خطای داخلی سرور
          schema:
            $ref: '#/definitions/Exchange_internal_model.ErrorResponseStruct'
      summary: ورود کاربر
      tags:
      - auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Revoke access token and end user session
      produces:
      - application/json
      responses:
        "200":
          description: Logout successful
          schema:
            $ref: '#/definitions/Exchange_internal_model.Response'
        "400":
          description: Bad request or missing authorization
          schema:
            $ref: '#/definitions/Exchange_internal_model.ErrorResponseStruct'
        "401":
          description: Unauthorized or invalid token
          schema:
            $ref: '#/definitions/Exchange_internal_model.ErrorResponseStruct'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/Exchange_internal_model.ErrorResponseStruct'
      security:
      - BearerAuth: []
      summary: User logout
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Generate new access token using refresh token
      parameters:
      - description: Refresh token info
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/Exchange_internal_model.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Token refreshed successfully
          schema:
            $ref: '#/definitions/Exchange_internal_model.Response'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/Exchange_internal_model.ErrorResponseStruct'
        "401":
          description: Unauthorized or expired token
          schema:
            $ref: '#/definitions/Exchange_internal_model.ErrorResponseStruct'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/Exchange_internal_model.ErrorResponseStruct'
      summary: Refresh access token
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: ایجاد حساب کاربری جدید با ایمیل، نام کامل، رمز عبور و اعتبارسنجی
        کپچا
      parameters:
      - description: اطلاعات ثبت‌نام شامل کپچا
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/Exchange_internal_model.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: ثبت‌نام موفق
          schema:
            $ref: '#/definitions/Exchange_internal_model.Response'
        "400":
          description: بدنه درخواست نامعتبر یا کپچا اشتباه
          schema:
            $ref: '#/definitions/Exchange_internal_model.ErrorResponseStruct'
        "409":
          description: کاربر قبلا ثبت شده
          schema:
            $ref: '#/definitions/Exchange_internal_model.ErrorResponseStruct'
        "500":
          description: خطای داخلی سرور
          schema:
            $ref: '#/definitions/Exchange_internal_model.ErrorResponseStruct'
      summary: ثبت نام کاربر
      tags:
      - auth
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"